refactor: extract shared adjacency BFS + WeaponProfile; tighten misc rejections (#21)
Picked the high-value items from the 3-agent reuse/quality/efficiency
sweep; deferred the bigger-scope refactors (denormalization migrations,
cross-tool boilerplate consolidation, CommandAckResponse unification,
MountDied cleanup applier) as follow-up work.

Done:

* **Shared adjacency BFS** — new `world/core/.../internal/movement/AdjacencyBfs.kt`
  with top-level `hopDistance` and `fleeCandidates`. AttackNpcReducer's
  private copies removed; both reducers now share it.
* **Shared WeaponProfile** — new `world/core/.../internal/combat/WeaponProfile.kt`
  with `weaponProfileFor` + `scalingEffectFor`. AttackNpcReducer's
  private copies removed.
* **TameReducer flee fix** — was single-hop only; now goes through the
  shared `fleeCandidates` with a new `balance.tameSpookFleeDistance()`
  (default 2). The `fleeCandidatesFromAttacker` helper with its
  `@Suppress("UNUSED_PARAMETER")` smell is gone.
* **Q4 rename** — `WorldRejection.IncompatibleMaintenanceResource.mount: MountId`
  → `target: UUID` to match the now-generic `maintain` verb (forward
  compat for `item:` / `building:` future variants).
* **Q2 sentinel killed** — TameReducer's `MOUNT_PLACEHOLDER = MountId(UUID(0L,0L))`
  hack for surfacing same-node-mismatch is gone. New `WorldRejection
  .TameTargetNotAtSameNode(agent, npc, agentAt, npcAt)` carries proper
  fields.
* **Q5 InspectTool through PrefixedIds** — `inspectMount` now routes
  through `PrefixedIds.parseMount` / `encodeMount` / `encodeAgent`
  instead of hand-rolled string handling.

Deferred (would be its own PR — too cross-cutting):
- Cross-tool boilerplate (`submitQueued`) + CommandAckResponse unification.
- MountDied cleanup consumer (orphan gear + cargo on death).
- Denormalize SADDLE / HARNESS / current-load on mount row to kill per-move
  reads.
- MountCargoServiceImpl 4 near-duplicate cargo methods.
- InspectTool's 3 sequential reads for mount inspect (batch).
- `MountInspectView` Perception-banded HP (design principle #7).
